server/particle/requirements.txt
influxdb
paho-mqtt
logging